pyproject.toml
[project]
name = "python-scan-4x4"
version = "0.1.0"
description = "A4 document scanner to 2x2 grid splitter - scan documents and split into 4 equal parts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pyobjc-framework-quartz>=12.1",
    "python-sane>=2.9.2",
    "requests>=2.32.5",
]
